Various changes and added features
Removed collision, step up on feature and overlap events on the darkness sprites.
Disabled the white paper sprite on the player as it was blocking the view.
Added Power Up Pop Up (Has a bug where it displays a white image when no power up is active)
Moved score from Game Mode to Game Instance