Added more scoring things. Changed timer
+10 Points when popping an empty bubble.
+ 50 Points when popping a bubble with an enemy.
+ 200 Points when picking up a power up
Fixed bug when adding timer to the score at the end of the level. Now it does it as soon as the level is completed. Not after the 5 second delay.
Linked bookcase letter bp to the letter spawner instead of the item bp.
Now the timer lasts 120s instead of 60s. Once the timer gets to 0 the player dies.
Substituted bubblegum bp for the general powerup bp in level 1 for consistency